+++ crates/tuirealm/src/core/subscription.rs
@@ -20,6 +20,7 @@ where
     U: Eq + PartialEq + Clone + PartialOrd,
 {
     /// Creates a new `Sub`
+    #[must_use]
     pub fn new(event_clause: EventClause<U>, sub_clause: SubClause<K>) -> Self {
         Self(event_clause, sub_clause)
     }
@@ -57,6 +58,7 @@ where
     U: Eq + PartialEq + Clone + PartialOrd + Send,
 {
     /// Instantiates a new [`Subscription`]
+    #[must_use]
     pub fn new(target: K, sub: Sub<K, U>) -> Self {
         Self {
             target,
@@ -66,16 +68,19 @@ where
     }
 
     /// Returns sub target
+    #[must_use]
     pub(crate) fn target(&self) -> &K {
         &self.target
     }
 
     /// Returns reference to subscription event clause
+    #[must_use]
     pub(crate) fn event(&self) -> &EventClause<U> {
         &self.ev
     }
 
     /// Returns whether to forward event to component
+    #[must_use]
     pub(crate) fn forward<HasAttrFn, GetStateFn, MountedFn>(
         &self,
         ev: &Event<U>,
@@ -190,8 +195,20 @@ where
     Not(Box<SubClause<Id>>),
     /// Forward event if both the inner clauses are `true`
     And(Box<SubClause<Id>>, Box<SubClause<Id>>),
+    /// Forward event if all the inner clauses are `true`.
+    ///
+    /// Short-circuits on first `false`.
+    ///
+    /// If empty will always return `false`.
+    AndMany(Vec<SubClause<Id>>),
     /// Forward event if at least one of the inner clauses is `true`
     Or(Box<SubClause<Id>>, Box<SubClause<Id>>),
+    /// Forward event if at least one of the inner clauses is `true`.
+    ///
+    /// Short-circuits on first `true`.
+    ///
+    /// If empty will always return `false`.
+    OrMany(Vec<SubClause<Id>>),
 }
 
 impl<Id> SubClause<Id>
@@ -200,21 +217,25 @@ where
 {
     /// Shortcut for [`SubClause::Not`] without specifying `Box::new(...)`
     #[allow(clippy::should_implement_trait)]
+    #[must_use]
     pub fn not(clause: Self) -> Self {
         Self::Not(Box::new(clause))
     }
 
     /// Shortcut for [`SubClause::And`] without specifying `Box::new(...)`
+    #[must_use]
     pub fn and(a: Self, b: Self) -> Self {
         Self::And(Box::new(a), Box::new(b))
     }
 
     /// Shortcut for [`SubClause::Or`] without specifying `Box::new(...)`
+    #[must_use]
     pub fn or(a: Self, b: Self) -> Self {
         Self::Or(Box::new(a), Box::new(b))
     }
 
     /// Returns whether the subscription clause is satisfied
+    #[must_use]
     pub(crate) fn forward<HasAttrFn, GetStateFn, MountedFn>(
         &self,
         has_attr_fn: HasAttrFn,
@@ -230,6 +251,8 @@ where
             .0
     }
 
+    /// Function to recursively check forwarding.
+    #[must_use]
     fn check_forwarding<HasAttrFn, GetStateFn, MountedFn>(
         &self,
         has_attr_fn: HasAttrFn,
@@ -267,6 +290,23 @@ where
                     b.check_forwarding(has_attr_fn, get_state_fn, mounted_fn);
                 (fwd_a && fwd_b, has_attr_fn, get_state_fn, mounted_fn)
             }
+            Self::AndMany(clauses) => {
+                let mut has_attr_fn = has_attr_fn;
+                let mut get_state_fn = get_state_fn;
+                let mut mounted_fn = mounted_fn;
+                for clause in clauses {
+                    let res = clause.check_forwarding(has_attr_fn, get_state_fn, mounted_fn);
+                    has_attr_fn = res.1;
+                    get_state_fn = res.2;
+                    mounted_fn = res.3;
+                    // short-circuit on any "false" value, which would correspond to "false && true && true" -> "false"
+                    if !res.0 {
+                        return (false, has_attr_fn, get_state_fn, mounted_fn);
+                    }
+                }
+                // return "false" on empty as anything else would not make sense to always forward if the condition is empty
+                (!clauses.is_empty(), has_attr_fn, get_state_fn, mounted_fn)
+            }
             Self::Or(a, b) => {
                 let (fwd_a, has_attr_fn, get_state_fn, mounted_fn) =
                     a.check_forwarding(has_attr_fn, get_state_fn, mounted_fn);
@@ -274,11 +314,28 @@ where
                     b.check_forwarding(has_attr_fn, get_state_fn, mounted_fn);
                 (fwd_a || fwd_b, has_attr_fn, get_state_fn, mounted_fn)
             }
+            Self::OrMany(clauses) => {
+                let mut has_attr_fn = has_attr_fn;
+                let mut get_state_fn = get_state_fn;
+                let mut mounted_fn = mounted_fn;
+                for clause in clauses {
+                    let res = clause.check_forwarding(has_attr_fn, get_state_fn, mounted_fn);
+                    has_attr_fn = res.1;
+                    get_state_fn = res.2;
+                    mounted_fn = res.3;
+                    // short-circuit on any "true" value, which would correspond to "false || true || true" -> "true"
+                    if res.0 {
+                        return (true, has_attr_fn, get_state_fn, mounted_fn);
+                    }
+                }
+                (false, has_attr_fn, get_state_fn, mounted_fn)
+            }
         }
     }
 
     // -- privates
 
+    #[must_use]
     fn has_attribute<HasAttrFn>(
         id: &Id,
         query: &Attribute,
@@ -297,6 +354,7 @@ where
         )
     }
 
+    #[must_use]
     fn has_state<GetStateFn>(id: &Id, state: &State, get_state_fn: GetStateFn) -> (bool, GetStateFn)
     where
         GetStateFn: Fn(&Id) -> Option<State>,
@@ -310,6 +368,7 @@ where
         )
     }
 
+    #[must_use]
     fn is_mounted<MountedFn>(id: &Id, mounted_fn: MountedFn) -> (bool, MountedFn)
     where
         MountedFn: Fn(&Id) -> bool,
@@ -320,7 +379,6 @@ where
 
 #[cfg(test)]
 mod test {
-
     use pretty_assertions::assert_eq;
 
     use super::*;
@@ -389,6 +447,208 @@ mod test {
         assert_eq!(
             sub.forward(
                 &Event::WindowResize(0, 0),
+                |_: &MockComponentId, q| component.query(q),
+                |_: &MockComponentId| Some(component.state()),
+                |_: &MockComponentId| true
+            ),
+            false
+        );
+    }
+
+    #[test]
+    fn forward_many() {
+        let ev: Event<MockEvent> = Event::Keyboard(Key::Char('q').into());
+        let mut component = MockFooInput::default();
+        component.attr(Attribute::Focus, AttrValue::Flag(true));
+
+        // AndMany all "true", returns "true"
+        let sub = Subscription::new(
+            MockComponentId::InputFoo,
+            Sub(
+                EventClause::Keyboard(Key::Char('q').into()),
+                SubClause::AndMany(vec![
+                    SubClause::IsMounted(MockComponentId::InputFoo),
+                    SubClause::IsMounted(MockComponentId::InputBar),
+                    SubClause::IsMounted(MockComponentId::InputOmar),
+                ]),
+            ),
+        );
+        assert_eq!(sub.target(), &MockComponentId::InputFoo);
+        assert_eq!(
+            sub.event(),
+            &EventClause::<MockEvent>::Keyboard(Key::Char('q').into())
+        );
+        assert_eq!(
+            sub.when,
+            SubClause::AndMany(vec![
+                SubClause::IsMounted(MockComponentId::InputFoo),
+                SubClause::IsMounted(MockComponentId::InputBar),
+                SubClause::IsMounted(MockComponentId::InputOmar),
+            ])
+        );
+        assert_eq!(
+            sub.forward(
+                &ev,
+                |_: &MockComponentId, q| component.query(q),
+                |_: &MockComponentId| Some(component.state()),
+                |_: &MockComponentId| true
+            ),
+            true
+        );
+
+        // AndMany one "false", returns "false"
+        let sub = Subscription::new(
+            MockComponentId::InputFoo,
+            Sub(
+                EventClause::Keyboard(Key::Char('q').into()),
+                SubClause::AndMany(vec![
+                    SubClause::IsMounted(MockComponentId::InputFoo),
+                    SubClause::IsMounted(MockComponentId::InputBar),
+                    SubClause::not(SubClause::IsMounted(MockComponentId::InputOmar)),
+                ]),
+            ),
+        );
+        assert_eq!(sub.target(), &MockComponentId::InputFoo);
+        assert_eq!(
+            sub.event(),
+            &EventClause::<MockEvent>::Keyboard(Key::Char('q').into())
+        );
+        assert_eq!(
+            sub.when,
+            SubClause::AndMany(vec![
+                SubClause::IsMounted(MockComponentId::InputFoo),
+                SubClause::IsMounted(MockComponentId::InputBar),
+                SubClause::not(SubClause::IsMounted(MockComponentId::InputOmar)),
+            ])
+        );
+        assert_eq!(
+            sub.forward(
+                &ev,
+                |_: &MockComponentId, q| component.query(q),
+                |_: &MockComponentId| Some(component.state()),
+                |_: &MockComponentId| true
+            ),
+            false
+        );
+
+        // OrMany one "false", returns "true"
+        let sub = Subscription::new(
+            MockComponentId::InputFoo,
+            Sub(
+                EventClause::Keyboard(Key::Char('q').into()),
+                SubClause::OrMany(vec![
+                    SubClause::IsMounted(MockComponentId::InputFoo),
+                    SubClause::IsMounted(MockComponentId::InputBar),
+                    SubClause::not(SubClause::IsMounted(MockComponentId::InputOmar)),
+                ]),
+            ),
+        );
+        assert_eq!(sub.target(), &MockComponentId::InputFoo);
+        assert_eq!(
+            sub.event(),
+            &EventClause::<MockEvent>::Keyboard(Key::Char('q').into())
+        );
+        assert_eq!(
+            sub.when,
+            SubClause::OrMany(vec![
+                SubClause::IsMounted(MockComponentId::InputFoo),
+                SubClause::IsMounted(MockComponentId::InputBar),
+                SubClause::not(SubClause::IsMounted(MockComponentId::InputOmar)),
+            ])
+        );
+        assert_eq!(
+            sub.forward(
+                &ev,
+                |_: &MockComponentId, q| component.query(q),
+                |_: &MockComponentId| Some(component.state()),
+                |_: &MockComponentId| true
+            ),
+            true
+        );
+
+        // OrMany all "false", returns "false"
+        let sub = Subscription::new(
+            MockComponentId::InputFoo,
+            Sub(
+                EventClause::Keyboard(Key::Char('q').into()),
+                SubClause::OrMany(vec![
+                    SubClause::not(SubClause::IsMounted(MockComponentId::InputFoo)),
+                    SubClause::not(SubClause::IsMounted(MockComponentId::InputBar)),
+                    SubClause::not(SubClause::IsMounted(MockComponentId::InputOmar)),
+                ]),
+            ),
+        );
+        assert_eq!(sub.target(), &MockComponentId::InputFoo);
+        assert_eq!(
+            sub.event(),
+            &EventClause::<MockEvent>::Keyboard(Key::Char('q').into())
+        );
+        assert_eq!(
+            sub.when,
+            SubClause::OrMany(vec![
+                SubClause::not(SubClause::IsMounted(MockComponentId::InputFoo)),
+                SubClause::not(SubClause::IsMounted(MockComponentId::InputBar)),
+                SubClause::not(SubClause::IsMounted(MockComponentId::InputOmar)),
+            ])
+        );
+        assert_eq!(
+            sub.forward(
+                &ev,
+                |_: &MockComponentId, q| component.query(q),
+                |_: &MockComponentId| Some(component.state()),
+                |_: &MockComponentId| true
+            ),
+            false
+        );
+    }
+
+    #[test]
+    fn forward_many_zero_elements() {
+        let ev: Event<MockEvent> = Event::Keyboard(Key::Char('q').into());
+        let mut component = MockFooInput::default();
+        component.attr(Attribute::Focus, AttrValue::Flag(true));
+
+        // AndMany returns "true"
+        let sub = Subscription::new(
+            MockComponentId::InputFoo,
+            Sub(
+                EventClause::Keyboard(Key::Char('q').into()),
+                SubClause::AndMany(vec![]),
+            ),
+        );
+        assert_eq!(sub.target(), &MockComponentId::InputFoo);
+        assert_eq!(
+            sub.event(),
+            &EventClause::<MockEvent>::Keyboard(Key::Char('q').into())
+        );
+        assert_eq!(sub.when, SubClause::AndMany(vec![]));
+        assert_eq!(
+            sub.forward(
+                &ev,
+                |_: &MockComponentId, q| component.query(q),
+                |_: &MockComponentId| Some(component.state()),
+                |_: &MockComponentId| true
+            ),
+            false
+        );
+
+        // OrMany returns "true"
+        let sub = Subscription::new(
+            MockComponentId::InputFoo,
+            Sub(
+                EventClause::Keyboard(Key::Char('q').into()),
+                SubClause::OrMany(vec![]),
+            ),
+        );
+        assert_eq!(sub.target(), &MockComponentId::InputFoo);
+        assert_eq!(
+            sub.event(),
+            &EventClause::<MockEvent>::Keyboard(Key::Char('q').into())
+        );
+        assert_eq!(sub.when, SubClause::OrMany(vec![]));
+        assert_eq!(
+            sub.forward(
+                &ev,
                 |_: &MockComponentId, q| component.query(q),
                 |_: &MockComponentId| Some(component.state()),
                 |_: &MockComponentId| true